+++ lib/smnoisebandpartition.hh
@@ -33,7 +33,7 @@ class NoiseBandPartition
 
 public:
   NoiseBandPartition (size_t n_bands, size_t n_spectrum_bins, double mix_freq);
-  void noise_envelope_to_spectrum (SpectMorph::Random& random_gen, const std::vector<double>& envelope, float *spectrum, double scale);
+  void noise_envelope_to_spectrum (SpectMorph::Random& random_gen, const std::vector<float>& envelope, float *spectrum, double scale);
 
   size_t n_bands();
   size_t n_spectrum_bins();
